git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene2621@1188713 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/src/java/org/apache/lucene/search/DisjunctionSumScorer.java
@@ -55,7 +55,7 @@ class DisjunctionSumScorer extends Scorer {
   /** The number of subscorers that provide the current match. */
   protected int nrMatchers = -1;
 
-  private float currentScore = Float.NaN;
+  private double currentScore = Float.NaN;
   
   /** Construct a <code>DisjunctionScorer</code>.
    * @param weight The weight to be used.
@@ -195,7 +195,7 @@ class DisjunctionSumScorer extends Scorer {
    * Initially invalid, until {@link #nextDoc()} is called the first time.
    */
   @Override
-  public float score() throws IOException { return currentScore; }
+  public float score() throws IOException { return (float)currentScore; }
    
   @Override
   public int docID() {
